Add parser options for sequencing platform.
Add parser option for sequencing platform.
Add parer option for structure-string.
Remove parser ooption for barcode length, as it is implicitly
included in platform or stucture string.
Add struct ReadStructure that hold the information on
the structure of a read pair and the position of barcode, sequence etc.
Started implementation of parsing function for the structure string (work in progress).